refactor: seperate particles to custom hook
