+++ src/App.tsx
@@ -1,36 +1,21 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { useEffect, useMemo, useState } from "react";
-import Particles, { initParticlesEngine } from "@tsparticles/react";
-import { loadSlim } from "@tsparticles/slim";
-import Login from "./pages/Login";
-import Dashboard from "./pages/Dashboard";
-import Error from "./pages/Error";
-import { particlesOptions } from "./utils/particles-config";
-import { Container } from "@tsparticles/engine";
+import Particles from "@tsparticles/react";
+import Login from "@pages/Login";
+import Dashboard from "@pages/Dashboard";
+import Error from "@pages/Error";
+import { useParticles } from "@hooks/useParticles";
 
 const router = createBrowserRouter([
-	{ path: "/", element: <Dashboard />, errorElement: <Error /> },
+	{
+		path: "/",
+		element: <Dashboard />,
+		errorElement: <Error />,
+	},
 	{ path: "/login", element: <Login /> },
 ]);
 
 function App() {
-	const [init, setInit] = useState(false);
-
-	useEffect(() => {
-		initParticlesEngine(async (engine) => {
-			await loadSlim(engine);
-		}).then(() => {
-			setInit(true);
-		});
-	}, []);
-
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	const particlesLoaded = (container: Container | undefined): Promise<void> => {
-		return new Promise<void>((resolve) => {
-			resolve();
-		});
-	};
-	const options = useMemo(() => particlesOptions, []);
+	const { init, particlesLoaded, options } = useParticles();
 
 	if (init) {
 		return (
@@ -43,8 +28,6 @@ function App() {
 				<RouterProvider router={router} />
 			</>
 		);
-	}
-
-	return <RouterProvider router={router} />;
+	} else return <RouterProvider router={router} />;
 }
 export default App;
